backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
httpx==0.28.1
chromadb==0.5.23
sentence-transformers==3.3.1
rank-bm25==0.2.2
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
Pillow==11.0.0
trafilatura==1.12.2
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.4.3
numpy==2.1.3
